tools/pyproject.toml
[project]
name = "deephaven-core-build"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0", "httpx>=0.27"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.16",
]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[project.scripts]
validate = "validate:main"
check-links = "check_links:main"
check-external-links = "check_external_links:main"
check-python = "check_python:main"
run-each-block = "run_each_block:main"
lint-blocks = "lint_blocks:main"
download-eval-data = "download_eval_data:main"
